+++ pkgs/unified_analytics/lib/src/analytics.dart
@@ -184,53 +184,6 @@ abstract class Analytics {
     );
   }
 
-  /// Factory constructor to return the [AnalyticsImpl] class with a
-  /// [MemoryFileSystem] to use for testing.
-  @visibleForTesting
-  factory Analytics.test({
-    required DashTool tool,
-    required Directory homeDirectory,
-    required String measurementId,
-    required String apiSecret,
-    required String dartVersion,
-    required FileSystem fs,
-    required DevicePlatform platform,
-    String? flutterChannel,
-    String? flutterVersion,
-    String? clientIde,
-    String? enabledFeatures,
-    SurveyHandler? surveyHandler,
-    GAClient? gaClient,
-    int toolsMessageVersion = kToolsMessageVersion,
-    String toolsMessage = kToolsMessage,
-  }) {
-    final firstRun = runInitialization(homeDirectory: homeDirectory, fs: fs);
-
-    return FakeAnalytics(
-      tool: tool,
-      homeDirectory: homeDirectory,
-      flutterChannel: flutterChannel,
-      toolsMessageVersion: toolsMessageVersion,
-      flutterVersion: flutterVersion,
-      dartVersion: dartVersion,
-      platform: platform,
-      fs: fs,
-      surveyHandler: surveyHandler ??
-          FakeSurveyHandler.fromList(
-            dismissedSurveyFile: fs.file(p.join(
-              homeDirectory.path,
-              kDartToolDirectoryName,
-              kDismissedSurveyFileName,
-            )),
-            initializedSurveys: [],
-          ),
-      gaClient: gaClient ?? const FakeGAClient(),
-      clientIde: clientIde,
-      enabledFeatures: enabledFeatures,
-      firstRun: firstRun,
-    );
-  }
-
   /// The shared identifier for Flutter and Dart related tooling using
   /// package:unified_analytics.
   String get clientId;
@@ -335,6 +288,53 @@ abstract class Analytics {
   ///
   /// The snooze period is defined by the [Survey.snoozeForMinutes] field.
   void surveyShown(Survey survey);
+
+  /// Returns an instance of [FakeAnalytics] which can be used in tests to check
+  /// for certain [Event] instances within [FakeAnalytics.sentEvents].
+  @visibleForTesting
+  static FakeAnalytics fake({
+    required DashTool tool,
+    required Directory homeDirectory,
+    required String dartVersion,
+    required MemoryFileSystem fs,
+    String? flutterChannel,
+    String? flutterVersion,
+    String? clientIde,
+    String? enabledFeatures,
+    SurveyHandler? surveyHandler,
+    GAClient? gaClient,
+    DevicePlatform platform = DevicePlatform.linux,
+    int toolsMessageVersion = kToolsMessageVersion,
+    String toolsMessage = kToolsMessage,
+    bool enableAsserts = true,
+  }) {
+    final firstRun = runInitialization(homeDirectory: homeDirectory, fs: fs);
+
+    return FakeAnalytics._(
+      tool: tool,
+      homeDirectory: homeDirectory,
+      flutterChannel: flutterChannel,
+      toolsMessageVersion: toolsMessageVersion,
+      flutterVersion: flutterVersion,
+      dartVersion: dartVersion,
+      platform: platform,
+      fs: fs,
+      surveyHandler: surveyHandler ??
+          FakeSurveyHandler.fromList(
+            dismissedSurveyFile: fs.file(p.join(
+              homeDirectory.path,
+              kDartToolDirectoryName,
+              kDismissedSurveyFileName,
+            )),
+            initializedSurveys: [],
+          ),
+      gaClient: gaClient ?? const FakeGAClient(),
+      clientIde: clientIde,
+      enabledFeatures: enabledFeatures,
+      firstRun: firstRun,
+      enableAsserts: enableAsserts,
+    );
+  }
 }
 
 class AnalyticsImpl implements Analytics {
@@ -765,7 +765,7 @@ class FakeAnalytics extends AnalyticsImpl {
   final List<Event> sentEvents = [];
 
   /// Class to use when you want to see which events were sent
-  FakeAnalytics({
+  FakeAnalytics._({
     required super.tool,
     required super.homeDirectory,
     required super.dartVersion,