+++ pkgs/unified_analytics/test/unified_analytics_test.dart
@@ -19,7 +19,7 @@ import 'package:unified_analytics/unified_analytics.dart';
 import 'package:yaml/yaml.dart';
 
 void main() {
-  late FileSystem fs;
+  late MemoryFileSystem fs;
   late Directory home;
   late Directory dartToolDirectory;
   late Analytics initializationAnalytics;
@@ -33,8 +33,6 @@ void main() {
   const homeDirName = 'home';
   const initialTool = DashTool.flutterTool;
   const secondTool = DashTool.dartTool;
-  const measurementId = 'measurementId';
-  const apiSecret = 'apiSecret';
   const toolsMessageVersion = 1;
   const toolsMessage = 'toolsMessage';
   const flutterChannel = 'flutterChannel';
@@ -55,11 +53,9 @@ void main() {
 
     // This is the first analytics instance that will be used to demonstrate
     // that events will not be sent with the first run of analytics
-    initializationAnalytics = Analytics.test(
+    initializationAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -77,11 +73,9 @@ void main() {
     //
     // This instance should have the same parameters as the one above for
     // [initializationAnalytics]
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -90,7 +84,7 @@ void main() {
       fs: fs,
       platform: platform,
       clientIde: clientIde,
-    ) as FakeAnalytics;
+    );
     analytics.clientShowedMessage();
 
     // The 5 files that should have been generated
@@ -165,11 +159,9 @@ void main() {
     // for session data
     sessionFile.writeAsStringSync('not a valid session id');
 
-    analytics = Analytics.test(
+    analytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -178,7 +170,7 @@ void main() {
       fs: fs,
       platform: platform,
       clientIde: clientIde,
-    ) as FakeAnalytics;
+    );
     analytics.clientShowedMessage();
 
     // Invoking a send command should reset the session file to a good state
@@ -229,11 +221,9 @@ void main() {
 
   test('New tool is successfully added to config file', () {
     // Create a new instance of the analytics class with the new tool
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: secondTool,
       homeDirectory: home,
-      measurementId: 'measurementId',
-      apiSecret: 'apiSecret',
       flutterChannel: 'ey-test-channel',
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -322,11 +312,9 @@ void main() {
 
     // Start up a second instance to simulate starting another
     // command being run
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -366,11 +354,9 @@ void main() {
 
     // Initialize a second analytics class, which simulates a second tool
     // Create a new instance of the analytics class with the new tool
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: secondTool,
       homeDirectory: home,
-      measurementId: 'measurementId',
-      apiSecret: 'apiSecret',
       flutterChannel: 'ey-test-channel',
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -390,11 +376,9 @@ void main() {
       'Two concurrent instances are running '
       'and reflect an accurate up to date telemetry status', () async {
     // Initialize a second analytics class, which simulates a second tool
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: secondTool,
       homeDirectory: home,
-      measurementId: 'measurementId',
-      apiSecret: 'apiSecret',
       flutterChannel: 'ey-test-channel',
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -444,11 +428,9 @@ void main() {
 
     // Initialize a second analytics class, which simulates a second tool
     // which should correct the missing trailing new line character
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: secondTool,
       homeDirectory: home,
-      measurementId: 'measurementId',
-      apiSecret: 'apiSecret',
       flutterChannel: 'ey-test-channel',
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,
@@ -475,11 +457,9 @@ void main() {
     // Initialize a second analytics class for the same tool as
     // the first analytics instance except with a newer version for
     // the tools message and version
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion + 1,
       toolsMessage: toolsMessage,
@@ -602,11 +582,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     //
     // This second instance should reset the config file when it goes
     // to increment the version in the file
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion + 1,
       toolsMessage: toolsMessage,
@@ -627,11 +605,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     // Creating a third instance after the second instance
     // has reset the config file should include the newly added
     // tool again with its incremented version number
-    final thirdAnalytics = Analytics.test(
+    final thirdAnalytics = Analytics.fake(
       tool: initialTool,
       homeDirectory: home,
-      measurementId: measurementId,
-      apiSecret: apiSecret,
       flutterChannel: flutterChannel,
       toolsMessageVersion: toolsMessageVersion + 1,
       toolsMessage: toolsMessage,
@@ -700,11 +676,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     // Set the clock to the start value defined above
     withClock(Clock.fixed(start), () {
       // This class will be constructed at a fixed time
-      final secondAnalytics = Analytics.test(
+      final secondAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         flutterChannel: flutterChannel,
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -729,11 +703,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
       // A new instance will need to be created since the second
       // instance in the previous block is scoped - this new instance
       // should not reset the files generated by the second instance
-      final thirdAnalytics = Analytics.test(
+      final thirdAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         flutterChannel: flutterChannel,
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -745,7 +717,7 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
       thirdAnalytics.clientShowedMessage();
 
       // Calling the send event method will result in the session file
-      // getting updated but because we use the `Analytics.test()` constructor
+      // getting updated but because we use the `Analytics.fake()` constructor
       // no events will be sent
       thirdAnalytics.send(testEvent);
 
@@ -773,11 +745,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     // Set the clock to the start value defined above
     withClock(Clock.fixed(start), () {
       // This class will be constructed at a fixed time
-      final secondAnalytics = Analytics.test(
+      final secondAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         flutterChannel: flutterChannel,
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -806,11 +776,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
       // A new instance will need to be created since the second
       // instance in the previous block is scoped - this new instance
       // should not reset the files generated by the second instance
-      final thirdAnalytics = Analytics.test(
+      final thirdAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         flutterChannel: flutterChannel,
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -822,7 +790,7 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
       thirdAnalytics.clientShowedMessage();
 
       // Calling the send event method will result in the session file
-      // getting updated but because we use the `Analytics.test()` constructor
+      // getting updated but because we use the `Analytics.fake()` constructor
       // no events will be sent
       thirdAnalytics.send(testEvent);
 
@@ -957,11 +925,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     // run for a given tool
     Analytics? secondAnalytics;
     for (var i = 0; i < 2; i++) {
-      secondAnalytics = Analytics.test(
+      secondAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         flutterChannel: flutterChannel,
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -1064,11 +1030,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
     // run for a given tool
     Analytics? secondAnalytics;
     for (var i = 0; i < 2; i++) {
-      secondAnalytics = Analytics.test(
+      secondAnalytics = Analytics.fake(
         tool: secondTool,
         homeDirectory: home,
-        measurementId: 'measurementId',
-        apiSecret: 'apiSecret',
         // flutterChannel: flutterChannel,  THIS NEEDS TO REMAIN REMOVED
         toolsMessageVersion: toolsMessageVersion,
         toolsMessage: toolsMessage,
@@ -1215,11 +1179,9 @@ Privacy Policy (https://policies.google.com/privacy).
   test('Consent message is formatted correctly for any tool other than flutter',
       () {
     // Create a new instance of the analytics class with the new tool
-    final secondAnalytics = Analytics.test(
+    final secondAnalytics = Analytics.fake(
       tool: secondTool,
       homeDirectory: home,
-      measurementId: 'measurementId',
-      apiSecret: 'apiSecret',
       flutterChannel: 'ey-test-channel',
       toolsMessageVersion: toolsMessageVersion,
       toolsMessage: toolsMessage,